tests/requirements.txt
pytest
pytest-mock
boto3
botocore
moto
